Finished my signal implementation, needs testing
Also did some cleaning up with typedefs, and fixed a compile warning in
my LibC

+++ clib/src/stdio/fbuf.c
@@ -87,7 +87,7 @@ int _ffill(FILE * stream, void *ptr, size_t size)
 		int retval = 0;
 
 		/* Determine if we read more than a full buffer */
-		if (size > fbuffer->Size) {
+		if (size > (size_t)fbuffer->Size) {
 			retval = Syscall4(MOLLENOS_SYSCALL_VFSREAD, MOLLENOS_SYSCALL_PARAM(stream->fd),
 				MOLLENOS_SYSCALL_PARAM(ptr), MOLLENOS_SYSCALL_PARAM(size),
 				MOLLENOS_SYSCALL_PARAM(&errcode));
@@ -119,7 +119,7 @@ int _ffill(FILE * stream, void *ptr, size_t size)
 
 		/* Sanity, if we read more than buffer 
 		 * we are still invalid */
-		if (size > fbuffer->Size) {
+		if (size > (size_t)fbuffer->Size) {
 			/* Invalidate */
 			fbuffer->Pointer = -1;
